Shortened flowchart name to just  "Flowchart.drawio.png"
Initially, I wanted to add "(Please download file for Visibility)" to let the reader know that they knew to download the raw file, open using images, and zoom in to look at the flowchart. I decided that it would be better to shorten the file name and just make a summary (which is just process flow diagram).